[QUAR-57] ♻️ Refactor: 소셜로그인 관련 코드 리팩토링
